requirements.txt
django == 5.0
djangorestframework == 3.14.0
psycopg2-binary == 2.9.9
drf-yasg == 1.21.7
python-dotenv == 1.0.0
coverage == 7.3.4